Move routing table specifics to golem common
