+++ golem-service-base/src/model.rs
@@ -3213,96 +3213,6 @@ impl Component {
     }
 }
 
-#[derive(Clone)]
-pub struct NumberOfShards {
-    pub value: usize,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct Pod {
-    host: String,
-    port: u16,
-}
-
-impl Pod {
-    pub fn uri(&self) -> Uri {
-        Uri::builder()
-            .scheme("http")
-            .authority(format!("{}:{}", self.host, self.port).as_str())
-            .path_and_query("/")
-            .build()
-            .expect("Failed to build URI")
-    }
-}
-
-impl From<GrpcPod> for Pod {
-    fn from(value: GrpcPod) -> Self {
-        Self {
-            host: value.host,
-            port: value.port as u16,
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct RoutingTable {
-    pub number_of_shards: NumberOfShards,
-    shard_assignments: HashMap<ShardId, Pod>,
-}
-
-impl RoutingTable {
-    pub fn lookup(&self, worker_id: &WorkerId) -> Option<&Pod> {
-        self.shard_assignments.get(&ShardId::from_worker_id(
-            &worker_id.clone().into(),
-            self.number_of_shards.value,
-        ))
-    }
-
-    pub fn random(&self) -> Option<&Pod> {
-        self.shard_assignments
-            .values()
-            .choose(&mut rand::thread_rng())
-    }
-
-    pub fn first(&self) -> Option<&Pod> {
-        self.shard_assignments.values().next()
-    }
-
-    pub fn all(&self) -> HashSet<&Pod> {
-        self.shard_assignments.values().collect()
-    }
-}
-
-impl From<GrpcRoutingTable> for RoutingTable {
-    fn from(value: GrpcRoutingTable) -> Self {
-        Self {
-            number_of_shards: NumberOfShards {
-                value: value.number_of_shards as usize,
-            },
-            shard_assignments: value
-                .shard_assignments
-                .into_iter()
-                .map(RoutingTableEntry::from)
-                .map(|routing_table_entry| (routing_table_entry.shard_id, routing_table_entry.pod))
-                .collect(),
-        }
-    }
-}
-
-pub struct RoutingTableEntry {
-    shard_id: ShardId,
-    pod: Pod,
-}
-
-impl From<GrpcRoutingTableEntry> for RoutingTableEntry {
-    fn from(value: GrpcRoutingTableEntry) -> Self {
-        Self {
-            shard_id: value.shard_id.unwrap().into(),
-            pod: value.pod.unwrap().into(),
-        }
-    }
-}
-
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Ord, PartialOrd, Serialize, Deserialize, Object)]
 #[serde(rename_all = "camelCase")]
 #[oai(rename_all = "camelCase")]
