+++ golem-service-base/src/routing_table.rs
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::model::RoutingTable;
 use async_trait::async_trait;
 use golem_api_grpc::proto::golem::shardmanager;
 use golem_api_grpc::proto::golem::shardmanager::shard_manager_service_client;
 use golem_common::cache::*;
 use serde::Deserialize;
 use url::Url;
+use golem_common::model::RoutingTable;
 
 #[derive(Debug, Clone)]
 pub enum RoutingTableError {
